requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
channels>=4.0.0
channels_redis>=4.1.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
openai>=1.0.0
stripe>=7.0.0
